torrent.php displays VIP and Trusted user
Setting a users "uploaderstatus" to 2 or 3 will set their status to either trusted (2) or VIP (3). This will result in an icon being placed next to their name on "torrent.php" indicating this user is "safe to download from"

The "verified user" images will need to be placed in the "css" folder as trusted.png and vip.png... Or you could just edit the code.