Added Torrent Tracker
Added a very lightweight torrent tracker written completely in PHP. This will allow users to make torrents with their own tracker and avoid the "0 peer/seeds" as it is processed as a https request and not udp which is likely to be blocked on most cheap webservice providers. No need to modify the database as "peers.txt" is used to connect peers to eachother. Strictly no torrent information is stored - only peer information to connect users to eachother when downloading.